+++ Source/EventFlow.TestHelpers/Aggregates/Queries/IDbContext.cs
@@ -1,0 +1,7 @@
+﻿namespace EventFlow.TestHelpers.Aggregates.Queries
+{
+    public interface IDbContext 
+    {
+        string Id { get; }
+    }
+}